# RideSync.app/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RideSync account, ride, booking, route, and passenger matching API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a RideSync account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created and session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to RideSync
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated session cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      operationId: getSession
      tags: [auth]
      summary: Load the current session
      responses:
        "200":
          description: Current authenticated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /rides/{id}/bookings:
    parameters:
      - $ref: "#/components/parameters/EntityId"
    get:
      operationId: listRideBookings
      tags: [bookings]
      summary: List booking requests for a ride owned by the current user
      responses:
        "200":
          description: Requests for this ride
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /bookings/{id}/status:
    parameters:
      - $ref: "#/components/parameters/EntityId"
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Approve or reject a request, or cancel the current user's booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusInput"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
    EntityId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 24
        maxLength: 24
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Record not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict with current record state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterInput:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 128
        college:
          type: string
          minLength: 2
          maxLength: 160
      required: [name, email, password, college]
    LoginInput:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 128
      required: [email, password]
    PublicUser:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        college: { type: string }
      required: [id, name, college]
    SessionUser:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        college: { type: string }
        memberSince: { type: string, format: date-time }
      required: [id, name, email, college, memberSince]
    SessionResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/SessionUser"
      required: [user]
...
    LocationInput:
      type: object
      additionalProperties: false
      properties:
        label: { type: string, minLength: 2, maxLength: 240 }
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
      required: [label, lat, lng]
...
    BookingStatusInput:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [confirmed, rejected, cancelled]
      required: [status]
    Booking:
      type: object
      properties:
        id: { type: string }
        rideId: { type: string }
        passenger: { $ref: "#/components/schemas/PublicUser" }
        ride:
          $ref: "#/components/schemas/BookingRideSummary"
        seats: { type: integer }
        status:
          type: string
          enum: [pending, confirmed, rejected, cancelled, completed]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, rideId, passenger, ride, seats, status, createdAt, updatedAt]
    BookingRideSummary:
      type: object
      properties:
        driver: { $ref: "#/components/schemas/PublicUser" }
        pickup: { $ref: "#/components/schemas/LocationInput" }
        destination: { $ref: "#/components/schemas/LocationInput" }
        departureAt: { type: string, format: date-time }
        availableSeats: { type: integer }
        contributionPerSeat: { type: number }
        status:
          type: string
          enum: [active, cancelled, completed]
      required: [driver, pickup, destination, departureAt, availableSeats, contributionPerSeat, status]
    BookingList:
      type: object
      properties:
        bookings:
          type: array
          items: { $ref: "#/components/schemas/Booking" }
      required: [bookings]
...